pyseries/requirements.txt
panda
matplotlib
numpy
tensorflow